Start to make command APIs clearer (#61)
* start on clearer protocols

* fix 3.7

* Update npe2/_types.py

* Update npe2/_types.py


* return to previous comment
