+++ npe2/_from_npe1.py
@@ -6,7 +6,6 @@ import warnings
 from configparser import ConfigParser
 from dataclasses import dataclass
 from functools import lru_cache
-from importlib.metadata import PackageNotFoundError
 from pathlib import Path
 from typing import (
     Any,
@@ -62,9 +61,6 @@ class HookSpecs:
 for m in dir(HookSpecs):
     if m.startswith("napari"):
         setattr(HookSpecs, m, napari_hook_specification(getattr(HookSpecs, m)))
-
-
-WidgetCallable = Union[Callable, Tuple[Callable, dict]]
 
 
 @dataclass
@@ -142,7 +138,7 @@ def norm_plugin_name(plugin_name: Optional[str] = None, module: Any = None) -> s
     msg = f"We tried hard! but could not detect a plugin named {plugin_name!r}."
     if plugin_manager.plugins:
         msg += f" Plugins found include: {list(plugin_manager.plugins)}"
-    raise PackageNotFoundError(msg)
+    raise metadata.PackageNotFoundError(msg)
 
 
 def manifest_from_npe1(
@@ -284,6 +280,7 @@ class HookImplParser:
                 warnings.warn(msg)
 
     def napari_experimental_provide_dock_widget(self, impl: HookImplementation):
+        WidgetCallable = Union[Callable, Tuple[Callable, dict]]
         items: Union[WidgetCallable, List[WidgetCallable]] = impl.function()
         if not isinstance(items, list):
             items = [items]  # pragma: no cover
